boot: zephyr: option to automatically upgrade
Add an option to automatically upgrade to valid images in the secondary
image slot. This is intended as a workaround for secondary images
where it is not possible to set the swap status from software due to
large write alignments.

This option can only be enabled with `BOOT_UPGRADE_ONLY` and
`MCUBOOT_DOWNGRADE_PREVENTION` to prevent it from being enabled in
normal use cases, where the standard SWAP behaviour is desired.

Signed-off-by: Jordan Yates <jordan@embeint.com>